Optimized GameObject multithreading

+++ src/main/scala-3/id/unikub/platFX/utils/transform/Transform.scala
@@ -1,3 +1,3 @@
 package id.unikub.platFX.utils.transform
 
-class Transform(var pos :Vector2 = Vector2(0, 0), var rot :Float = 0, var scl :Vector2 = Vector2(0, 0))
+class Transform(var pos :Vector2 = Vector2(0, 0), var rot :Double = 0, var scl :Vector2 = Vector2(0, 0))
